Add progress bar for uploads
This doesn't really work as expected because blazer buffers the file to
calculate the shasum before uploading so the progress bar fills to 100%
almost immediately and then hangs until the upload is actually complete.
I also tried this with several other libraries all with the same effect.
However, I'm committing this anyway in case I decide to rip out blazer
and use something else or write something else.